+++ code/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" >
+    xmlns:tools="http://schemas.android.com/tools">
 
     <application
         android:allowBackup="true"
@@ -11,19 +11,25 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.EventGate"
-        tools:targetApi="31" >
+        tools:targetApi="31">
         <activity
-            android:name=".admin.AdminActivity"
+            android:name=".Organizer.OrganizerCreateEventActivity"
             android:exported="false" />
         <activity
-            android:name=".organizer.OrganizerActivity"
+            android:name=".OrganizerMainMenuActivity"
             android:exported="false" />
         <activity
-            android:name=".attendee.AttendeeActivity"
+            android:name=".Admin.AdminActivity"
+            android:exported="false" />
+        <activity
+            android:name=".Organizer.OrganizerMainMenuActivity"
+            android:exported="false" />
+        <activity
+            android:name=".Attendee.AttendeeActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
-            android:exported="true" >
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
