+++ code/app/src/main/java/com/example/eventgate/Admin/AdminActivity.java
@@ -1,4 +1,4 @@
-package com.example.eventgate.admin;
+package com.example.eventgate.Admin;
 
 import androidx.appcompat.app.AppCompatActivity;
