+++ code/app/src/main/java/com/example/eventgate/Attendee/AttendeeActivity.java
@@ -1,4 +1,4 @@
-package com.example.eventgate.organizer;
+package com.example.eventgate.Attendee;
 
 import android.os.Bundle;
 
@@ -6,10 +6,10 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.eventgate.R;
 
-public class OrganizerActivity extends AppCompatActivity {
+public class AttendeeActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_organizer);
+        setContentView(R.layout.activity_attendee);
     }
 }
