--- code/app/src/main/java/com/example/eventgate/attendee/AttendeeActivity.java
@@ -1,15 +1,0 @@
-package com.example.eventgate.attendee;
-
-import android.os.Bundle;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import com.example.eventgate.R;
-
-public class AttendeeActivity extends AppCompatActivity {
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_attendee);
-    }
-}
