+++ code/app/src/main/res/values/strings.xml
@@ -1,3 +1,6 @@
 <resources>
     <string name="app_name">EventGate</string>
+    <string name="your_events">Your Events</string>
+    <string name="create_new_event">Create New Event</string>
+    <string name="back">Back</string>
 </resources>
